refactor(ai): extract registerAiTasks to its own file and clarify Phase 4 JSDoc
Per code review on Phase 4 commit d20d7e9:
- Extract registerAiTasks to packages/ai/src/task/registerAiTasks.ts so
  the audit test no longer imports from a banned barrel ("./index"). The
  index module re-exports the function so the public surface is unchanged.
- Clarify the JSDoc on plain-Task subclasses (RerankerTask,
  QueryExpanderTask, ModelSearchTask) so future readers know `requires`
  on these classes is informational only — they implement their own
  execute() and bypass AiTask.gateOrThrow. The audit test still validates
  the values are known capabilities.

Tests: 44/44 across @workglow/ai. Build green for @workglow/ai (wider
build remains red for chrome-ai + 8 other vendor packages — Phase 5).
