+++ packages/ai/src/task/RerankerTask.ts
@@ -118,7 +118,12 @@ interface RankedItem {
  */
 export class RerankerTask extends Task<RerankerTaskInput, RerankerTaskOutput, RerankerTaskConfig> {
   public static override type = "RerankerTask";
-  /** Capabilities required; used by the audit test to verify coverage. */
+  /**
+   * Informational: capability this task uses. NOT enforced by the dispatcher —
+   * RerankerTask extends `Task` (not `AiTask`) and implements its own `execute()`,
+   * so `gateOrThrow` is never called against this value. The audit test in
+   * `task/index.test.ts` validates the value is a known {@link Capability}.
+   */
   public static readonly requires: readonly Capability[] = ["text.reranking"] as const satisfies readonly Capability[];
   public static override category = "RAG";
   public static override title = "Reranker";
